change assigned for allocate
