--- ereuse_devicehub/resources/assigned/definitions.py
@@ -1,9 +1,0 @@
-from typing import Callable, Iterable, Tuple
-from ereuse_devicehub.resources.action import schemas
-from teal.resource import Resource
-from ereuse_devicehub.resources.assigned.views import AssignedView
-
-
-class AssignedDef(Resource):
-    VIEW = AssignedView
-    SCHEMA = schemas.Assigned
